--- app/src/main/java/com/jammin/myapplication/data/network/ServerAPI.kt
@@ -1,3 +1,0 @@
-package com.jammin.myapplication.data.network
-
-interface ServerAPI
